Extracting serveStatic() to a dedicated file.
