+++ src/index.ts
@@ -13,7 +13,8 @@ export { createLogger } from "./logger";
 export { createMiddleware } from "./middleware";
 export { createResultHandler, defaultResultHandler } from "./result-handler";
 export { DependsOnMethod } from "./depends-on-method";
-export { Routing, serveStatic } from "./routing";
+export { serveStatic } from "./serve-static";
+export { Routing } from "./routing";
 export { createServer, attachRouting } from "./server";
 export { OpenAPI } from "./open-api";
 export { OpenAPIError, DependsOnMethodError, RoutingError } from "./errors";
