+++ tests/unit/serve-static.spec.ts
@@ -1,0 +1,9 @@
+import { serveStatic } from "../../src";
+
+describe("serveStatic()", () => {
+  test("should return static handling function", () => {
+    expect(typeof serveStatic(__dirname, { dotfiles: "deny" })).toBe(
+      "function"
+    );
+  });
+});
